remove old axiellservicetest class, rename current
